now testing Pundit auth

+++ app/controllers/bookmarks_controller.rb
@@ -13,7 +13,7 @@ class BookmarksController < ApplicationController
     @bookmark = Bookmark.new(params.require(:bookmark).permit(:url))
     @bookmark.topic = @topic
 
-    @bookmark.authorize
+    authorize @bookmark
 
     if @bookmark.save
       redirect_to [@topic, @bookmark], notice: "Bookmark was saved successfully."
@@ -31,7 +31,7 @@ class BookmarksController < ApplicationController
     @bookmark = Bookmark.find(params[:id])
     @bookmark.url = params[:bookmark][:url]
 
-    @bookmark.authorize
+    authorize @bookmark
 
     if @bookmark.save
       redirect_to [@bookmark.topic, @bookmark], notice: "Bookmark was saved successfully."
@@ -44,7 +44,8 @@ class BookmarksController < ApplicationController
   def destroy
     @bookmark = Bookmark.find(params[:id])
 
-    @bookmark.authorize
+    authorize @bookmark
+
 
     if @bookmark.destroy
       flash[:notice] = "\"#{@bookmark.url}\" was deleted successfully."
